Added Request Approval mechanism for ResourceDepotController

+++ src/main/java/com/spm/resqjeevanredis/dto/ResourceInfoDto.java
@@ -11,6 +11,6 @@ public class ResourceInfoDto {
     private String resourceId;
     private String resourceName;
     private ResourceType resourceType;
-    private int unitsAvailable;
+    private long unitsAvailable;
     private String resourceDepotId;
 }
